Add list roles list groups tasks
Move all management tools in the new rbac/tools namespace
Consolidate similar tasks into a single file
